+++ staging/InquiryChamber/index.ts
@@ -1,0 +1,1 @@
+export { InquiryChamber, type InquiryChamberProps } from './InquiryChamber'