+++ staging/ShearSplit/index.ts
@@ -1,0 +1,1 @@
+export { ShearSplit, type ShearSplitProps } from './ShearSplit'